+++ src/K4os.Compression.LZ4.Test/Adapters/BaselineLZ4.cs
@@ -1,5 +1,4 @@
 using System;
-using Xunit;
 
 namespace K4os.Compression.LZ4.Test.Adapters
 {