+++ CHANGES.md
@@ -1,5 +1,5 @@
 ## 1.2.10-beta (2021/01/29)
-* ADDED #14: support for ReadOnlySequence<byte> and IBufferWriter<byte> (pickle only)
+* ADDED #14 (partial): support for IBufferWriter in LZ4Pickler
 
 ## 1.2.8-beta (2020/09/26)
 * Implicit reference to System.Runtime.CompilerServices.Unsafe